Boton de enviar en el form de gestion de peliculas anadido

+++ project_final_attempt/Entities/Peliculas/Movie.cs
@@ -12,18 +12,18 @@ namespace project_final_attempt.Entities.Peliculas
         private string nombre_pelicula;
         private int? Ano_presentacion;
         private List<string> directores;
-        private int? monto_recaudado;
+        private float? monto_recaudado;
         private string universo_perteneciente;
         private List<Personaje> personajes_pelicula;
 
         public string _name { get { return nombre_pelicula; } set { nombre_pelicula = value; } }
         public int? _year { get { return Ano_presentacion; } set { Ano_presentacion = value; } }
         public List<string> _directors { get { return directores; } set { directores = value; } }
-        public int? _amount { get { return monto_recaudado; } set { monto_recaudado = value; } }
+        public float? _amount { get { return monto_recaudado; } set { monto_recaudado = value; } }
         public string _universe { get { return universo_perteneciente; } set { universo_perteneciente = value; } }
         public List<Personaje> _casting { get { return personajes_pelicula; } set { personajes_pelicula = value; } }
 
-        public Movie(string name, int? year, List<string> directors_, int? amount, string universe, List<Personaje> casting)
+        public Movie(string name, int? year, List<string> directors_, float? amount, string universe, List<Personaje> casting)
         {
             _name = name;
             _year = year;
